Move isMessageRead and outlookAnchorFromItem into the policy module
